+++ pynwis/__init__.py
@@ -37,7 +37,7 @@ from .parameters import (
     search_parameters
 )
 
-__version__ = '0.1.0'
+__version__ = '0.1.1'
 __all__ = [
     'fetch_usgs_daily',
     'usgs_json_to_df',